requirements.txt
langchain==0.2.0
langchain-openai
langchain-community
openai
python-dotenv
pandas
tqdm
openpyxl
matplotlib
seaborn
anthropic
langchain-google-genai